Unify both branches of the liveness checker's mainloop
Not only is it (very) slightly smaller, but if the liveness checker
received a packet when it had no outstanding jobs it would go into a
tight loop of calling select without ever processing the packet.  That
should be extremely unlikely in a production system, but it's very
easy in development.

Reviewed-by: mark.slemko@socrata.com