Add basic auth in HttpClientHttpClient
In many backend services, our http client is tied to precisely ony
AuthScope and we wish to have BasicAuth against that host.  We could
set the header's and such on every request at the calling site, but I
think it would be cleaner to have this be set on the Apache HttpClient.

This commits creates an HttpRequestInterceptor that is added to the
HttpClient that uses the provided credentials provider and
well does basic auth.