Fix liveness checking in the presence of multiple checks for one target
This was ultimately caused by java.util.Map's `get` and `remove` methods
taking `Object` instead of `K`.  The reason we're using `ju.Map` here
instead of a Scala `Map` is because in other places in this code we're
using a Java `ConcurrentHashMap` and I thought it was important to have
just one map API in play.  Sigh.  Anyway I've now checked all uses of
`pings` to make sure the right keys are being used.